# pythoncode/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Library Management API",
    "description": "API for managing books in a library"
  },
  "host": "127.0.0.1:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/books": {
      "get": {
        "summary": "Get all books",
        "description": "Retrieve a list of all books",
        "responses": {
          "200": {
            "description": "A list of books",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "title": {"type": "string"},
                  "author": {"type": "string"},
                  "isbn": {"type": "string"},
                  "published_year": {"type": "integer"},
                  "genre": {"type": "string"}
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add books",
        "description": "Add a single book or multiple books",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {"type": "string"},
                "author": {"type": "string"},
                "isbn": {"type": "string"},
                "published_year": {"type": "integer"},
                "genre": {"type": "string"}
              }
            }
          }
        ],
        "responses": {
          "201": {"description": "Books added successfully"}
        }
      }
    },
    "/books/<isbn>": {
      "delete": {
        "summary": "Delete a book by ISBN",
        "description": "Delete a book using its ISBN",
        "parameters": [
          {"name": "isbn", "in": "path", "required": true, "type": "string"}
        ],
        "responses": {
          "200": {"description": "Book deleted successfully"},
          "404": {"description": "Book not found"}
        }
      },
      "put": {
        "summary": "Update book details by ISBN",
        "description": "Update a book's details using its ISBN",
        "parameters": [
          {"name": "isbn", "in": "path", "required": true, "type": "string"},
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {"type": "string"},
                "author": {"type": "string"},
                "isbn": {"type": "string"},
                "published_year": {"type": "integer"},
                "genre": {"type": "string"}
              }
            }
          }
        ],
        "responses": {
          "200": {"description": "Book updated successfully"},
          "404": {"description": "Book not found"}
        }
      }
    },
    "/books/search": {
      "get": {
        "summary": "Search books",
        "description": "Search books by author, published year, or genre",
        "parameters": [
          {"name": "author", "in": "query", "type": "string"},
          {"name": "published_year", "in": "query", "type": "integer"},
          {"name": "genre", "in": "query", "type": "string"}
        ],
        "responses": {
          "200": {
            "description": "A list of matching books",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "title": {"type": "string"},
                  "author": {"type": "string"},
                  "isbn": {"type": "string"},
                  "published_year": {"type": "integer"},
                  "genre": {"type": "string"}
                }
              }
            }
          }
        }
      }
    }
  }
}